Move source files to directory
