+++ metabolomics_spectrum_resolver/tasks.py
@@ -8,8 +8,7 @@ import joblib
 import redis
 import spectrum_utils.spectrum as sus
 
-import parsing
-import drawing
+from metabolomics_spectrum_resolver import drawing, parsing
 
 
 memory = joblib.Memory("tmp/joblibcache", verbose=0)
