+++ metabolomics_spectrum_resolver/templates/drawing_controls.html
