+++ run_worker.sh
@@ -3,4 +3,4 @@ source activate usi
 
 export C_FORCE_ROOT="true"
 #TODO: Make sure we don't run this worker as root
-celery -A tasks worker -l info --autoscale=10,1 -Q worker --max-tasks-per-child 10 --loglevel INFO
+celery -A metabolomics_spectrum_resolver.tasks worker -l info --autoscale=10,1 -Q worker --max-tasks-per-child 10 --loglevel INFO
